+++ src/main/java/graveldb/datastore/KeyValueStore.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 
 public interface KeyValueStore {
 
-    void put(String key, String value) throws IOException;
-    String get(String key) throws IOException;
-    void delete(String key) throws IOException;
+    void put(String key, String value);
+    String get(String key);
+    void delete(String key);
 
 }