+++ src/main/java/graveldb/datastore/bloomfilter/BloomFilter.java
@@ -10,29 +10,33 @@ import java.util.BitSet;
 
 import com.google.common.hash.Hashing;
 
-public class BloomFilterImpl {
+public class BloomFilter {
 
-    private static final Logger log = LoggerFactory.getLogger(BloomFilterImpl.class);
+    private static final Logger log = LoggerFactory.getLogger(BloomFilter.class);
 
     private static final int HASH_FUNCTIONS = 7;
-    private static final int BLOOM_BUCKET = 500000;
+    private static final int BLOOM_BUCKET = 5000000;
     private BitSet bitArray;
 
     private final String fileName;
 
-    public BloomFilterImpl(String fileName) throws IOException {
+    public BloomFilter(String fileName) {
         this.fileName = fileName;
         this.bitArray = new BitSet(BLOOM_BUCKET);
         Path path = Path.of(fileName);
-        if (!Files.exists(path)) {
-            Files.createDirectories(path.getParent());
-            Files.createFile(path);
-        } else generateBitArray();
+            if (!Files.exists(path)) {
+                try {
+                    Files.createDirectories(path.getParent());
+                    Files.createFile(path);
+                }catch (Exception e) {
+                    throw new RuntimeException(e);
+                }
+            } else generateBitArray();
     }
 
-    public BloomFilterWriter getWriter() throws FileNotFoundException { return new BloomFilterImpl.BloomFilterWriter(); }
+    public BloomFilterWriter getWriter() { return new BloomFilter.BloomFilterWriter(); }
 
-    public boolean check(String key) throws IOException {
+    public boolean check(String key) {
         for (int i=0; i<HASH_FUNCTIONS; i++) {
             int setBit = Math.abs(getKeyHashValue(key, i) % BLOOM_BUCKET);
             if (!checkSetBit(setBit)) return false;
@@ -40,20 +44,20 @@ public class BloomFilterImpl {
         return true;
     }
 
-    private boolean checkSetBit(int setBit) throws IOException {
+    private boolean checkSetBit(int setBit) {
         if (bitArray == null) generateBitArray();
         return bitArray.get(setBit);
     }
 
-    private void generateBitArray() throws IOException {
+    private void generateBitArray() {
         byte[] byteArray = new byte[(BLOOM_BUCKET + 7) / 8];
-        RandomAccessFile fis = new RandomAccessFile(fileName, "r");
-        try (fis) {
+
+        try (RandomAccessFile fis = new RandomAccessFile(fileName, "r");) {
             int res = fis.read(byteArray);
             if (res >= 0) bitArray = BitSet.valueOf(byteArray);
             else bitArray = new BitSet(BLOOM_BUCKET);
         } catch (IOException e) {
-            throw new IOException(e);
+            throw new RuntimeException(e);
         }
     }
 
@@ -61,8 +65,12 @@ public class BloomFilterImpl {
 
         BufferedOutputStream bos;
 
-        public BloomFilterWriter() throws FileNotFoundException {
-            bos = new BufferedOutputStream(new FileOutputStream(fileName));
+        public BloomFilterWriter() {
+            try {
+                bos = new BufferedOutputStream(new FileOutputStream(fileName));
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
         }
 
         public void write(String key) {